Added setcode for "Pendulum Dragon"
Now that the leaks have begun, we can already see that "Pendulum Dragon" is now going to be an official archetype. The logical step would be to put it at 0x1f0, since it's the first new code avaiable, but I was wondering: wouldn't it be best to put it as 0x11f0 and leave 0x1f0 as a super-series for cards like "Fusion Dragon", "Synchro Dragon", etc., just in case they're needed later? What do you think?