Added SPFE-JP005 --捕食植物サンデウ・キンジー
From what I understand, this card cannot be easily scripted: similar to "Fusion Tag" back in SHVI, Ygopro doesn't have functions like IsFusionAttribute() or effects like EFFECT_ADD_FUSION_ATTRIBUTE at the moment (or EFFECT_CHANGE_FUSION_ATTRIBUTE, in this case: the wording of the (1) effect implies that it has to be treated as DARK, it can't be used as the original Attribute anymore). Until those functions are added to the core, I thought something like this would work: it would flag cards affected by this effect, and then the fusion procedures of monsters that require certain attributes would check this as well. It would require altering all monsters that use determined attributes as Fusion Materials (at the moment I remember the manga E-HEROes, the El-Shaddolls, Starve Venom and Chimera Rafflesia), but it would simulate the effect until the core is updated. What do you think?