Damage Step Fix for Vermillion Dragon Mech
The effect reads,

    If this Synchro Summoned card is destroyed by card effect and sent to the Graveyard: You can...

I'm new to YGOPRO's scripting standards, but EVENT_TO_GRAVE cannot miss timing, correct? If it can then according to constant.lua, if I translated it correctly, adding EFFECT_FLAG_DELAY is necessary. Correct me if I'm wrong, but the effect doesn't exclude the Damage Step, so that's necessary to add as well...