Added Hint for Speedrift
Another card with a Special Summon effect for the End Phase. You must control only a single Tuner monster, so it's a little less likely to happen.

I added quite a few hints recently, where they're not really needed, more for people who forgot to hold the 'A' button in the End Phase, so if you disagree with adding them, kindly notify me and I'll close this request to allow other people's requests to advance.