Added End Phase Hint for Dinomist Howling
To remind the user to activate in the End Phase seems kind of useful when the restriction lasts so long, "you cannot Pendulum Summon monsters until the end of the next turn, except "Dinomist" monsters."

Therefore, if the game reminds you to activate, the sooner you activate it the faster that restriction will end, if that makes sense? It also serves as a way to "thin your deck", because when your turn starts, your deck will have up to 2 less cards in it, increasing the odds for you to draw the cards you want.