Added Hints for MACR cards (#87)
* Added End Phase Hint for Performapal Sky Pupil

For the Special Summon Quick Effect, like most of those effects that have this hint.

* Added Draw Phase Hint for Dark Contract with the Eternal Darkness

To aid the lock down effect, similar to "True King V.F.D., The Beast"'s:

* Added End Phase Hint for Destruction Swordsman's Cradle

For the Special Summon Quick Effect, like most of those effects that have this hint.

* Added End Phase Hints for Ethereal Wyrmterfall

Similar to other Quick Effect searchers, like Gleipnir, the Fetters of Fenrir, https://github.com/Fluorohydride/ygopro-scripts/blob/master/c14464864.lua#L7

The second End Phase Hint is there to remind the user to activate, just in case the opponent set a few negate cards that turn, if you activate Ethereal Wyrmterfall in the opponent's End Phase, they won't be able to negate it, because they just set their negate cards.

* Added End Phase Hint for Dinomist Howling

To remind the user to activate in the End Phase seems kind of useful when the restriction lasts so long, "you cannot Pendulum Summon monsters until the end of the next turn, except "Dinomist" monsters."

Therefore, if the game reminds you to activate, the sooner you activate it the faster that restriction will end, if that makes sense? It also serves as a way to "thin your deck", because when your turn starts, your deck will have up to 2 less cards in it, increasing the odds for you to draw the cards you want.