New procedure for Rock-Paper-Scissors
The original version with Duel.SelectOption() had the side-effect of showing your selection to the opponent before they made their choice, making the game basically doomed to failure. The new version, on the other hand, uses three tokens in the database, with a picture of rock, paper and scissors, and makes the players choose using Group.Select(). This offers better control over when to show the result, and has the additional bonus of resembling a RPS interface more than the old version.